PicoScope: Support T=double
This is needed for the MVP, as load_grc instantiates everything with
T=double.

Also unify the settings interface to use double for offsets/ranges/thresholds,
to avoid confusion.

+++ blocklib/picoscope/Picoscope4000a.hpp
@@ -367,7 +367,7 @@ struct Picoscope4000a : public fair::picoscope::Picoscope<T, Picoscope4000a<T>> 
             const auto coupling = detail::convertToPs4000aCoupling(channel.settings.coupling);
             const auto range    = detail::convertToPs4000aRange(channel.settings.range);
 
-            status              = ps4000aSetChannel(this->state.handle, *idx, true, coupling, static_cast<PICO_CONNECT_PROBE_RANGE>(range), channel.settings.offset);
+            status              = ps4000aSetChannel(this->state.handle, *idx, true, coupling, static_cast<PICO_CONNECT_PROBE_RANGE>(range), static_cast<float>(channel.settings.offset));
             if (status != PICO_OK) {
                 fmt::println(std::cerr, "ps4000aSetChannel (chan '{}'): {}", channel.id, detail::getErrorMessage(status));
                 return { status };
